added permission config

+++ app/Livewire/AccountForm.php
@@ -3,16 +3,21 @@
 namespace App\Livewire;
 
 use App\Models\Account;
+use Illuminate\Support\Facades\Gate;
+use Livewire\Attributes\Validate;
 use Livewire\Component;
 
 class AccountForm extends Component
 {
+    #[Validate('required|unique:accounts,name')]
     public $name;
     public $account;
 
     public $listeners = ['restore' => 'restore'];
 
     public function mount($id = null){
+        if(!Gate::allows('config-permission',3))
+            abort('404');
         if ($id) {
             $this->account = Account::findOrFail($id);
             $this->name = $this->account->name;
@@ -22,6 +27,9 @@ class AccountForm extends Component
     }
 
     public function save(){
+        if(!Gate::allows('config-permission',3))
+            abort('404');
+        $this->validate();
         $t = Account::where('name',$this->name)->withTrashed()->first();
         if ($t !=null) {
             if($t->trashed()){
@@ -40,18 +48,24 @@ class AccountForm extends Component
     }
 
     public function remove(){
+        if(!Gate::allows('config-permission',3))
+            abort('404');
         if($this->account->id == 1) return redirect()->route('dashboard.settings');
         $this->account->delete();
         return redirect()->route('dashboard.settings');
     }
 
     public function restore(){
+        if(!Gate::allows('config-permission',3))
+            abort('404');
         Account::where('name',$this->name)->restore();
         return redirect()->route('dashboard.settings');
     }
 
     public function render()
     {
+        if(!Gate::allows('config-permission',3))
+            abort('404');
         return view('livewire.account-form');
     }
 }
