+++ app/Livewire/CategoryForm.php
@@ -3,6 +3,7 @@
 namespace App\Livewire;
 
 use App\Models\Category;
+use Illuminate\Support\Facades\Gate;
 use Livewire\Attributes\Locked;
 use Livewire\Attributes\Validate;
 use Livewire\Component;
@@ -20,6 +21,8 @@ class CategoryForm extends Component
     public Category $category;
 
     public function mount($id = null){
+        if(!Gate::allows('config-permission',3))
+            abort('404');
         try {
             $this->category = Category::where('id',$id)->firstOrFail();
             $this->alias = $this->category->alias;
@@ -30,6 +33,8 @@ class CategoryForm extends Component
     }
 
     public function save(){
+        if(!Gate::allows('config-permission',3))
+            abort('404');
         $this->validate();
         $this->category->name = $this->name;
         $this->category->alias = $this->alias;
@@ -38,12 +43,16 @@ class CategoryForm extends Component
     }
 
     public function remove(){
+        if(!Gate::allows('config-permission',3))
+            abort('404');
         $this->category->delete();
         return $this->redirect(route('dashboard.settings'));
     }
 
     public function render()
     {
+        if(!Gate::allows('config-permission',3))
+            abort('404');
         return view('livewire.category-form');
     }
 }
