+++ app/Livewire/Userform.php
@@ -5,6 +5,7 @@ namespace App\Livewire;
 use App\Models\Person;
 use App\Models\User;
 use App\Models\UserPermission;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Validation\Rule;
 use Livewire\Attributes\Validate;
@@ -31,11 +32,19 @@ class Userform extends Component
     public $p4 = 1;
     public $p5 = 1;
     public $p6 = 1;
+    public $p7 = 1;
+    public $p8 = 1;
+    public $p9 = 1;
+    public $p10 = 1;
+    public $p11 = 1;
+    public $p12 = 1;
 
     public User $user;
     public UserPermission $permissions;
 
     public function mount($id = null){
+        if(!Gate::allows('config-permission',3))
+            abort('404');
         try {
             $this->user = User::findOrFail($id);
 
@@ -50,11 +59,17 @@ class Userform extends Component
                 $this->permissions = $this->user->permission;
                 $this->user->id= $this->user->permission->user_id;
                 $this->p1 = $this->user->permission->supplier;
-                $this->p2 = $this->user->permission->item;
-                $this->p3 = $this->user->permission->partner;
-                $this->p4 = $this->user->permission->product;
-                $this->p5 = $this->user->permission->delivery;
+                $this->p2 = $this->user->permission->product;
+                $this->p3 = $this->user->permission->item;
+                $this->p4 = $this->user->permission->delivery;
+                $this->p5 = $this->user->permission->partner;
                 $this->p6 = $this->user->permission->transaction;
+                $this->p7 = $this->user->permission->client;
+                $this->p8 = $this->user->permission->voucher;
+                $this->p9 = $this->user->permission->ledger;
+                $this->p10 = $this->user->permission->report;
+                $this->p11 = $this->user->permission->config;
+                $this->p12 = $this->user->permission->history;
             }
         } catch (\Throwable) {
             $this->user = new User();
@@ -63,6 +78,8 @@ class Userform extends Component
     }
 
     public function save(){
+        if(!Gate::allows('config-permission',3))
+            abort('404');
         Validator::make([
             'name' => $this->name,
             'email' => $this->email,
@@ -87,22 +104,32 @@ class Userform extends Component
 
         $this->permissions->user_id = $this->user->id ?? null;
         $this->permissions->supplier =  $this->p1;
-        $this->permissions->item = $this->p2;
-        $this->permissions->partner = $this->p3;
-        $this->permissions->product = $this->p4;
-        $this->permissions->delivery = $this->p5;
+        $this->permissions->product = $this->p2;
+        $this->permissions->item = $this->p3;
+        $this->permissions->delivery = $this->p4;
+        $this->permissions->partner = $this->p5;
         $this->permissions->transaction = $this->p6;
+        $this->permissions->client = $this->p7;
+        $this->permissions->voucher = $this->p8;
+        $this->permissions->ledger = $this->p9;
+        $this->permissions->report = $this->p10;
+        $this->permissions->config = $this->p11;
+        $this->permissions->history = $this->p12;
         $this->permissions->save();
         $this->redirect(route('dashboard.settings'));
     }
 
     public function remove(){
+        if(!Gate::allows('config-permission',3))
+            abort('404');
         $this->user->delete();
         $this->redirect(route('dashboard.settings'));
     }
 
     public function render()
     {
+        if(!Gate::allows('config-permission',3))
+            abort('404');
         $persons = Person::all();
         return view('livewire.userform',compact(['persons']));
     }
